Speed up imports
