+++ openai/api_resources/embedding.py
@@ -1,10 +1,10 @@
 import base64
 import time
 
-
 from openai import util
 from openai.api_resources.abstract.engine_api_resource import EngineAPIResource
-from openai.datalib import numpy as np, assert_has_numpy
+from openai.datalib.numpy_helper import assert_has_numpy
+from openai.datalib.numpy_helper import numpy as np
 from openai.error import TryAgain
